pyproject.toml
[build-system]
requires = [
  "setuptools>=42",
  "wheel"
]

build-backend = "setuptools.build_meta"

[project]
name = "exp_heatmap"
version = "1.3.0"
authors = [
    {name = "Edvard Ehler", email = "edvard.ehler@img.cas.cz"},
    {name = "Ondřej Moravčík"}
]
maintainers = [
    {name = "Adam Nógell", email = "adam.nogell@img.cas.cz"}
]
description = "Ordered-pair workflow for regional cross-population genomic visualization (VCF -> empirical rank-score heatmaps)"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = [
    "bioinformatics",
    "population-genetics",
    "selection-scan",
    "xpehh",
    "xpnsl",
    "fst",
    "tajima-d",
    "heatmap",
    "visualization",
    "1000-genomes",
    "ggvp"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Visualization"
]
dependencies = [
    "scikit-allel>=1.3.7",
    "zarr>=2.10,<3.0.0",
    "numpy>=1.22",
    "pandas>=1.4",
    "matplotlib>=3.5",
    "seaborn>=0.12",
    "click>=8.0",
    "plotly>=5.10",
    "tqdm>=4.60"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0"
]
benchmarks = [
    "psutil>=5.9",
    "pysam>=0.22"
]
all = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "psutil>=5.9",
    "pysam>=0.22"
]

[project.urls]
Homepage = "https://github.com/bioinfocz/exp_heatmap"
Repository = "https://github.com/bioinfocz/exp_heatmap"
"Bug Tracker" = "https://github.com/bioinfocz/exp_heatmap/issues"
Documentation = "https://github.com/bioinfocz/exp_heatmap#readme"

[project.scripts]
exp_heatmap = "exp_heatmap.cli:cli"
exp-heatmap = "exp_heatmap.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
